Test that pyproject dependencies can be built by conda
