+++ pyproject.toml
@@ -54,15 +54,15 @@ dependencies = [
     "auth0-python==4.0.0",
     "azure-identity==1.6.1",
     "azure-mgmt-containerservice==16.2.0",
-    "bcrypt==4.0.1",
-    "boto3==1.26.54",
+    "bcrypt==3.2.2",
+    "boto3==1.26.74",
     "cloudflare==2.11.1",
-    "kubernetes==25.3.0",
-    "pydantic==1.10.4",
+    "kubernetes==26.1.0",
+    "pydantic==1.10.5",
     "pynacl==1.5.0",
-    "python-keycloak==2.9.0",
+    "python-keycloak==2.12.0",
     "questionary==1.10.0",
-    "rich==13.2.0",
+    "rich==12.6.0",
     "ruamel.yaml==0.17.21",
     "typer==0.7.0",
 ]
