+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+markers =
+    conda: conda required to run this test (deselect with '-m \"not conda\"')
