+++ tests/test_dependencies.py
@@ -6,26 +6,12 @@ from pathlib import Path
 import pytest
 import ruamel.yaml
 
+from .conftest import TEST_CONDA_ENV_NAME
+
 PYPROJECT = Path(__file__).parent.parent / "pyproject.toml"
-CONDA_ENV_NAME = "nebari-test-dependencies"
 
 
-@pytest.fixture
-def conda_env_file(tmp_path):
-    """Create a temporary environment.yaml file."""
-
-    yield tmp_path / "environment.yaml"
-
-    # remove created conda environment
-    try:
-        subprocess.run(
-            ["conda", "env", "remove", "-n", CONDA_ENV_NAME, "-y"],
-            check=True,
-        )
-    except subprocess.CalledProcessError as e:
-        raise e
-
-
+@pytest.mark.conda
 def test_pyproject_dependencies(conda_env_file):
     """
     Populate a temporary environment.yaml, then attempt to create a conda environment with it.
@@ -33,7 +19,11 @@ def test_pyproject_dependencies(conda_env_file):
     This test ensures that nebari can be packaged by and available on conda-forge.
     """
 
-    template = {"name": CONDA_ENV_NAME, "channels": ["conda-forge"], "dependencies": []}
+    template = {
+        "name": TEST_CONDA_ENV_NAME,
+        "channels": ["conda-forge"],
+        "dependencies": [],
+    }
 
     assert PYPROJECT.exists()
 
@@ -41,9 +31,19 @@ def test_pyproject_dependencies(conda_env_file):
         data = tomllib.load(f)
 
     for dep in data["project"].get("dependencies"):
-        # split into package, constraint, version
-        p, c, v = re.split(r"([=><]{1,2})", dep)
-        template["dependencies"].append(f"{p} {c}{v}")
+        dep_details = re.split(r"([=><]{1,2})", dep)
+        if len(dep_details) > 1:
+            # split into package, constraint, version
+            p, c, v = dep_details
+        else:
+            p = dep_details[0]
+            c, v = "", ""
+
+        # kubernetes package on conda-forge is prepended with "python-"
+        if p == "kubernetes":
+            p = "python-" + p
+
+        template["dependencies"].append(f"{p}{c}{v}")
 
     with open(conda_env_file, "w+") as f:
         yaml = ruamel.yaml.YAML()
@@ -55,5 +55,6 @@ def test_pyproject_dependencies(conda_env_file):
             ["conda", "env", "create", "-f", conda_env_file],
             check=True,
         )
+
     except subprocess.CalledProcessError as e:
         raise e
